# Refactor EncryptParameterAssignmentToken with JDK library