--- a/shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-core/src/main/java/org/apache/shardingsphere/encrypt/rewrite/token/pojo/EncryptParameterAssignmentToken.java
+++ b/shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-core/src/main/java/org/apache/shardingsphere/encrypt/rewrite/token/pojo/EncryptParameterAssignmentToken.java
@@ -17,10 +17,9 @@
 
 package org.apache.shardingsphere.encrypt.rewrite.token.pojo;
 
-import com.google.common.collect.Collections2;
-
 import java.util.Collection;
 import java.util.LinkedList;
+import java.util.StringJoiner;
 
 /**
  * Parameter assignment token for encrypt.
@@ -44,7 +43,10 @@ public final class EncryptParameterAssignmentToken extends EncryptAssignmentToke
     
     @Override
     public String toString() {
-        Collection<String> items = Collections2.transform(columnNames, input -> input + " = ?");
-        return String.join(", ", items);
+        StringJoiner result = new StringJoiner(", ");
+        for (String input : columnNames) {
+            result.add(input + " = ?");
+        }
+        return result.toString();
     }
 }