Fixes #17687.

### Summary

```
Before:
EncryptParameterAssignmentTokenBenchmark.bench         thrpt   15  3113.274 ± 26.429  ops/ms

After:
EncryptParameterAssignmentTokenBenchmark.bench         thrpt   15  13868.939 ± 178.703  ops/ms
```

### Before


```
# JMH version: 1.33
# VM version: JDK 17.0.1, Java HotSpot(TM) 64-Bit Server VM, 17.0.1+12-LTS-39
# VM invoker: /usr/local/java/jdk-17.0.1/bin/java
# VM options: -XX:+SegmentedCodeCache -XX:+AggressiveHeap
# Blackhole mode: full + dont-inline hint (default, use -Djmh.blackhole.autoDetect=true to auto-detect)
# Warmup: 5 iterations, 3 s each
# Measurement: 5 iterations, 3 s each
# Timeout: 10 min per iteration
# Threads: 8 threads, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench

# Run progress: 0.00% complete, ETA 00:01:30
# Fork: 1 of 3
# Preparing profilers: AsyncProfiler 
# Warmup Iteration   1: 2717.733 ops/ms
# Warmup Iteration   2: 3132.016 ops/ms
# Warmup Iteration   3: 3074.729 ops/ms
# Warmup Iteration   4: 3137.834 ops/ms
# Warmup Iteration   5: 3152.314 ops/ms
Iteration   1: 3138.202 ops/ms
Iteration   2: 3091.051 ops/ms
Iteration   3: 3130.710 ops/ms
Iteration   4: 3036.444 ops/ms
Iteration   5: [WARN] Unknown argument: summary
3119.229 ops/ms
                 ·async: (text only)

# Processing profiler results: AsyncProfiler 

# Run progress: 33.33% complete, ETA 00:01:01
# Fork: 2 of 3
# Preparing profilers: AsyncProfiler 
# Warmup Iteration   1: 2873.575 ops/ms
# Warmup Iteration   2: 3120.873 ops/ms
# Warmup Iteration   3: 3097.420 ops/ms
# Warmup Iteration   4: 3124.724 ops/ms
# Warmup Iteration   5: 3089.387 ops/ms
Iteration   1: 3121.724 ops/ms
Iteration   2: 3123.199 ops/ms
Iteration   3: 3121.065 ops/ms
Iteration   4: 3108.972 ops/ms
Iteration   5: [WARN] Unknown argument: summary
3125.970 ops/ms
                 ·async: (text only)

# Processing profiler results: AsyncProfiler 

# Run progress: 66.67% complete, ETA 00:00:30
# Fork: 3 of 3
# Preparing profilers: AsyncProfiler 
# Warmup Iteration   1: 2888.200 ops/ms
# Warmup Iteration   2: 3085.853 ops/ms
# Warmup Iteration   3: 3123.679 ops/ms
# Warmup Iteration   4: 3135.261 ops/ms
# Warmup Iteration   5: 3098.285 ops/ms
Iteration   1: 3129.475 ops/ms
Iteration   2: 3102.560 ops/ms
Iteration   3: 3123.368 ops/ms
Iteration   4: 3125.997 ops/ms
Iteration   5: [WARN] Unknown argument: summary
3101.138 ops/ms
                 ·async: (text only)

# Processing profiler results: AsyncProfiler 


Result "icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench":
  3113.274 ±(99.9%) 26.429 ops/ms [Average]
  (min, avg, max) = (3036.444, 3113.274, 3138.202), stdev = 24.722
  CI (99.9%): [3086.845, 3139.703] (assumes normal distribution)

Secondary result "icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench:·async":
--- Execution profile ---
Total samples       : 11743
unknown_Java        : 776 (6.61%)
not_walkable_Java   : 1 (0.01%)

          ns  percent  samples  top
  ----------  -------  -------  ---
 11784521144   10.00%     1175  java.util.Formatter.format
  8641669214    7.33%      860  java.lang.AbstractStringBuilder.ensureCapacityInternal
  7357883813    6.25%      734  java.util.ArrayList.add
  7012762110    5.95%      697  java.lang.StringLatin1.indexOf
  6683570924    5.67%      666  jbyte_disjoint_arraycopy
  5401537104    4.58%      539  com.google.common.base.Joiner.appendTo
  4867458078    4.13%      486  java.util.Formatter.parse
  4641170062    3.94%      464  java.lang.StringBuilder.append
  4279421030    3.63%      427  java.util.Formatter$Flags.contains
  4004845144    3.40%      399  java.lang.String.length
  3870866632    3.29%      387  java.lang.AbstractStringBuilder.append
  3650166103    3.10%      365  java.util.ArrayList.grow
  2841623561    2.41%      284  org.apache.shardingsphere.encrypt.rewrite.token.pojo.EncryptParameterAssignmentToken.toString
  2731834899    2.32%      272  java.util.ArrayList.elementData
  2532504798    2.15%      252  java.util.Formatter$FormatSpecifier.<init>
  2081678124    1.77%      207  java.util.LinkedList.listIterator
  2031755323    1.72%      201  java.lang.CharacterDataLatin1.isUpperCase
  1922044601    1.63%      191  java.util.LinkedList$ListItr.next
  1801168035    1.53%      180  java.util.Formatter.toString
  1740240224    1.48%      174  com.google.common.collect.Collections2$TransformedCollection.iterator
  1651954654    1.40%      164  java.lang.String.isLatin1
  1600451564    1.36%      159  java.lang.Object.<init>
  1512461696    1.28%      151  com.google.common.collect.Iterators$6.transform
  1432027471    1.22%      143  jlong_disjoint_arraycopy
  1351089415    1.15%      135  java.util.Formatter.<init>
  1241102194    1.05%      124  java.lang.String.format
  1131198294    0.96%      113  com.google.common.collect.TransformedIterator.<init>
  1100957209    0.93%      109  com.google.common.base.Preconditions.checkNotNull
  1070570512    0.91%      107  org.apache.shardingsphere.encrypt.rewrite.token.pojo.EncryptParameterAssignmentToken.lambda$toString$0
   890687799    0.76%       88  java.lang.AbstractStringBuilder.newCapacity
   870552663    0.74%       87  java.lang.String.getBytes
   868797992    0.74%       86  com.google.common.base.Joiner.toString
   861032404    0.73%       85  java.util.Formatter$Conversion.isValid
   820979835    0.70%       82  java.lang.String.charAt
   799996398    0.68%       80  jdk.internal.util.ArraysSupport.newLength
   770740856    0.65%       77  jbyte_arraycopy
   669819430    0.57%       67  java.util.LinkedList$ListItr.hasNext
   620274104    0.53%       62  java.lang.AbstractStringBuilder.putStringAt
   580208242    0.49%       58  com.google.common.collect.TransformedIterator.next
   530070678    0.45%       53  MemAllocator::Allocation::notify_allocation_jfr_sampler()
   410258213    0.35%       41  java.lang.Character.isUpperCase
   390773843    0.33%       39  java.util.LinkedList.checkPositionIndex
   389930473    0.33%       39  java.lang.StringLatin1.charAt
   381090667    0.32%       38  java.util.Arrays.copyOf
   331339525    0.28%       33  com.google.common.collect.Iterators$6.<init>
   299274274    0.25%       30  ObjAllocator::initialize(HeapWordImpl**) const
   270451931    0.23%       27  java.util.LinkedList.isPositionIndex
   260576878    0.22%       26  org.openjdk.jmh.infra.Blackhole.consumeFull
   259996728    0.22%       26  java.lang.String.indexOf
   259942481    0.22%       26  ObjArrayAllocator::initialize(HeapWordImpl**) const
Async profiler results:
  /home/wuweijie/projects/database-jmh/icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench-Throughput/summary-cpu.txt
--- Execution profile ---
Total samples       : 11763
unknown_Java        : 594 (5.05%)
skipped             : 1 (0.01%)

          ns  percent  samples  top
  ----------  -------  -------  ---
 16053080513   13.61%     1604  java.util.ArrayList.add
 15713816269   13.32%     1566  java.util.Formatter.format
  6751981135    5.72%      674  java.util.ArrayList.get
  6494407889    5.51%      646  java.util.Formatter.parse
  6125909324    5.19%      611  jbyte_disjoint_arraycopy
  5490486001    4.66%      549  icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench
  4660727455    3.95%      464  java.util.LinkedList$ListItr.next
  4001264370    3.39%      399  org.apache.shardingsphere.encrypt.rewrite.token.pojo.EncryptParameterAssignmentToken.lambda$toString$0
  3911887106    3.32%      389  java.util.Formatter$FormatSpecifier.print
  3612125285    3.06%      360  java.util.ArrayList.grow
  3331530960    2.82%      333  java.lang.String.getBytes
  2531562288    2.15%      253  com.google.common.base.Joiner.join
  2412467988    2.05%      241  java.lang.AbstractStringBuilder.append
  2241459315    1.90%      224  java.lang.StringLatin1.indexOf
  2240652425    1.90%      224  java.lang.AbstractStringBuilder.ensureCapacityInternal
  2061396718    1.75%      203  java.util.Formatter$FormatSpecifier.printString
  1800712217    1.53%      180  java.util.Objects.checkIndex
  1669676208    1.42%      167  com.google.common.collect.TransformedIterator.next
  1562107663    1.32%      156  java.lang.String.length
  1521282050    1.29%      152  java.lang.AbstractStringBuilder.appendChars
  1481646616    1.26%      148  jlong_disjoint_arraycopy
  1360761669    1.15%      136  com.google.common.collect.TransformedIterator.<init>
  1311395068    1.11%      131  com.google.common.collect.Iterators$6.transform
  1059300442    0.90%      106  java.util.Formatter$FormatSpecifier.index
   940043502    0.80%       94  com.google.common.collect.Iterators.transform
   910486576    0.77%       91  java.lang.String.format
   861481321    0.73%       85  java.lang.StringBuilder.append
   809662169    0.69%       81  org.apache.shardingsphere.encrypt.rewrite.token.pojo.EncryptParameterAssignmentToken$$Lambda$41.0x0000000800c2c4c0.apply
   800783968    0.68%       80  java.util.Formatter$Conversion.isValid
   790823695    0.67%       79  com.google.common.collect.TransformedIterator.hasNext
   771173739    0.65%       77  jbyte_arraycopy
   770474798    0.65%       76  java.util.Formatter$Flags.contains
   750655043    0.64%       75  com.google.common.base.Joiner.appendTo
   650410797    0.55%       65  java.lang.AbstractStringBuilder.checkRange
   630919350    0.53%       63  java.util.LinkedList$ListItr.<init>
   580844527    0.49%       57  java.util.LinkedList$ListItr.hasNext
   500487337    0.42%       50  MemAllocator::Allocation::notify_allocation_jfr_sampler()
   440371690    0.37%       44  java.util.Formatter.<init>
   430669800    0.37%       43  java.util.Formatter$FixedString.print
   410529259    0.35%       41  java.lang.CharacterDataLatin1.isUpperCase
   400442922    0.34%       40  com.google.common.collect.Collections2$TransformedCollection.iterator
   370244487    0.31%       36  java.lang.AbstractStringBuilder.putStringAt
   370230606    0.31%       37  ObjArrayAllocator::initialize(HeapWordImpl**) const
   349861875    0.30%       35  java.lang.String.coder
   330291774    0.28%       33  com.google.common.base.Preconditions.checkNotNull
   320366772    0.27%       32  ObjAllocator::initialize(HeapWordImpl**) const
   300287310    0.25%       30  java.util.Formatter$FixedString.<init>
   289664110    0.25%       29  java.util.Formatter.ensureOpen
   240218481    0.20%       24  java.lang.AbstractStringBuilder.getCoder
   240163940    0.20%       24  org.openjdk.jmh.infra.Blackhole.consumeFull
Async profiler results:
  /home/wuweijie/projects/database-jmh/icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench-Throughput/summary-cpu.txt
--- Execution profile ---
Total samples       : 11755
unknown_Java        : 591 (5.03%)

          ns  percent  samples  top
  ----------  -------  -------  ---
 22150308257   18.79%     2210  java.util.Formatter.format
 15373771513   13.04%     1534  java.util.ArrayList.add
  6723325806    5.70%      671  java.util.Formatter.parse
  6716147837    5.70%      668  jbyte_disjoint_arraycopy
  5481740269    4.65%      546  icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench
  4422008947    3.75%      441  java.util.LinkedList$ListItr.next
  3841063172    3.26%      384  org.apache.shardingsphere.encrypt.rewrite.token.pojo.EncryptParameterAssignmentToken.lambda$toString$0
  3612785381    3.06%      361  java.util.ArrayList.grow
  3141197633    2.66%      314  java.lang.String.getBytes
  3101373960    2.63%      307  java.util.Formatter$FormatSpecifier.print
  2673585479    2.27%      267  java.util.Formatter$FormatSpecifier.printString
  2462533790    2.09%      245  com.google.common.base.Joiner.join
  2261524435    1.92%      226  java.lang.StringLatin1.indexOf
  1940454542    1.65%      194  java.util.Formatter.<init>
  1883751704    1.60%      188  com.google.common.collect.TransformedIterator.<init>
  1731770288    1.47%      173  java.lang.AbstractStringBuilder.ensureCapacityInternal
  1660972422    1.41%      166  java.util.Formatter$FixedString.print
  1541178858    1.31%      154  java.lang.String.length
  1522002995    1.29%      152  com.google.common.collect.TransformedIterator.next
  1521826440    1.29%      152  java.lang.AbstractStringBuilder.appendChars
  1459735283    1.24%      146  java.util.Formatter$Flags.contains
  1341724476    1.14%      134  com.google.common.collect.Iterators$6.transform
  1301376877    1.10%      130  jlong_disjoint_arraycopy
  1239934609    1.05%      124  java.util.Objects.checkIndex
  1121569894    0.95%      112  java.lang.AbstractStringBuilder.append
  1011016656    0.86%       99  org.apache.shardingsphere.encrypt.rewrite.token.pojo.EncryptParameterAssignmentToken$$Lambda$37.0x0000000800c2ce40.apply
   920681490    0.78%       92  java.lang.StringBuilder.append
   871279781    0.74%       87  java.util.Formatter$Conversion.isValid
   790283858    0.67%       79  com.google.common.collect.TransformedIterator.hasNext
   780050839    0.66%       78  com.google.common.collect.Iterators.transform
   650048696    0.55%       65  jbyte_arraycopy
   620631101    0.53%       62  com.google.common.base.Joiner.appendTo
   620381030    0.53%       60  MemAllocator::Allocation::notify_allocation_jfr_sampler()
   590669196    0.50%       59  java.lang.CharacterDataLatin1.isUpperCase
   560847625    0.48%       56  java.lang.AbstractStringBuilder.checkRange
   489784852    0.42%       49  java.util.LinkedList$ListItr.checkForComodification
   489445533    0.42%       48  java.util.ArrayList.elementData
   480558649    0.41%       48  java.lang.String.format
   441487523    0.37%       44  java.lang.AbstractStringBuilder.putStringAt
   429959628    0.36%       43  java.util.LinkedList$ListItr.<init>
   389869899    0.33%       39  java.util.Arrays.copyOf
   370195280    0.31%       37  java.util.Formatter$FormatSpecifier.index
   331086038    0.28%       33  java.util.Formatter.ensureOpen
   320458295    0.27%       32  java.lang.StringLatin1.charAt
   310193749    0.26%       31  com.google.common.collect.Iterators$6.<init>
   280098314    0.24%       28  ObjAllocator::initialize(HeapWordImpl**) const
   260692994    0.22%       26  com.google.common.base.Preconditions.checkNotNull
   260302223    0.22%       26  java.util.LinkedList$ListItr.hasNext
   259804307    0.22%       26  com.google.common.collect.Collections2$TransformedCollection.iterator
   240271062    0.20%       23  java.lang.String.coder
Async profiler results:
  /home/wuweijie/projects/database-jmh/icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench-Throughput/summary-cpu.txt


# Run complete. Total time: 00:01:31

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

Benchmark                                               Mode  Cnt     Score    Error   Units
EncryptParameterAssignmentTokenBenchmark.bench         thrpt   15  3113.274 ± 26.429  ops/ms
EncryptParameterAssignmentTokenBenchmark.bench:·async  thrpt            NaN              ---

```


### After

```
# JMH version: 1.33
# VM version: JDK 17.0.1, Java HotSpot(TM) 64-Bit Server VM, 17.0.1+12-LTS-39
# VM invoker: /usr/local/java/jdk-17.0.1/bin/java
# VM options: -XX:+SegmentedCodeCache -XX:+AggressiveHeap
# Blackhole mode: full + dont-inline hint (default, use -Djmh.blackhole.autoDetect=true to auto-detect)
# Warmup: 5 iterations, 3 s each
# Measurement: 5 iterations, 3 s each
# Timeout: 10 min per iteration
# Threads: 8 threads, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench

# Run progress: 0.00% complete, ETA 00:01:30
# Fork: 1 of 3
# Preparing profilers: AsyncProfiler 
# Warmup Iteration   1: 11857.156 ops/ms
# Warmup Iteration   2: 13081.091 ops/ms
# Warmup Iteration   3: 13055.990 ops/ms
# Warmup Iteration   4: 12875.185 ops/ms
# Warmup Iteration   5: 14040.790 ops/ms
Iteration   1: 13836.124 ops/ms
Iteration   2: 14016.284 ops/ms
Iteration   3: 13975.880 ops/ms
Iteration   4: 13854.622 ops/ms
Iteration   5: [WARN] Unknown argument: summary
14040.379 ops/ms
                 ·async: (text only)

# Processing profiler results: AsyncProfiler 

# Run progress: 33.33% complete, ETA 00:01:01
# Fork: 2 of 3
# Preparing profilers: AsyncProfiler 
# Warmup Iteration   1: 13159.989 ops/ms
# Warmup Iteration   2: 13865.879 ops/ms
# Warmup Iteration   3: 14002.878 ops/ms
# Warmup Iteration   4: 13742.966 ops/ms
# Warmup Iteration   5: 13862.933 ops/ms
Iteration   1: 13723.511 ops/ms
Iteration   2: 14011.851 ops/ms
Iteration   3: 13819.349 ops/ms
Iteration   4: 13703.142 ops/ms
Iteration   5: [WARN] Unknown argument: summary
13618.464 ops/ms
                 ·async: (text only)

# Processing profiler results: AsyncProfiler 

# Run progress: 66.67% complete, ETA 00:00:30
# Fork: 3 of 3
# Preparing profilers: AsyncProfiler 
# Warmup Iteration   1: 13136.095 ops/ms
# Warmup Iteration   2: 13953.217 ops/ms
# Warmup Iteration   3: 13810.267 ops/ms
# Warmup Iteration   4: 13924.420 ops/ms
# Warmup Iteration   5: 13921.736 ops/ms
Iteration   1: 13497.128 ops/ms
Iteration   2: 13998.156 ops/ms
Iteration   3: 13956.974 ops/ms
Iteration   4: 14008.003 ops/ms
Iteration   5: [WARN] Unknown argument: summary
13974.212 ops/ms
                 ·async: (text only)

# Processing profiler results: AsyncProfiler 


Result "icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench":
  13868.939 ±(99.9%) 178.703 ops/ms [Average]
  (min, avg, max) = (13497.128, 13868.939, 14040.379), stdev = 167.159
  CI (99.9%): [13690.235, 14047.642] (assumes normal distribution)

Secondary result "icu.wwj.jmh.shardingsphere5.code.EncryptParameterAssignmentTokenBenchmark.bench:·async":
--- Execution profile ---
Total samples       : 11774
unknown_Java        : 10 (0.08%)

          ns  percent  samples  top
  ----------  -------  -------  ---
 24534216900   20.79%     2449  java.lang.String.length
 21286705280   18.04%     2120  jbyte_disjoint_arraycopy
 19982596763   16.94%     1995  org.apache.shardingsphere.encrypt.rewrite.token.pojo.EncryptParameterAssignmentToken.toString
 10617180311    9.00%     1060  java.lang.String.getBytes
  7844180707    6.65%      779  java.util.LinkedList$ListItr.next
  5851967010    4.96%      585  java.lang.String.join
  4930099002    4.18%      493  java.util.StringJoiner.checkAddLength
  4782199100    4.05%      478  java.lang.String.coder
  4273237706    3.62%      427  java.util.StringJoiner.add
  3671842337    3.11%      367  jbyte_arraycopy
  2643080298    2.24%      264  java.lang.String.valueOf
   850750013    0.72%       85  java.util.LinkedList$ListItr.hasNext
   810442793    0.69%       81  org.openjdk.jmh.infra.Blackhole.consumeFull
   780087919    0.66%       78  jdk.internal.misc.Unsafe.allocateUninitializedArray
   720317773    0.61%       72  jint_disjoint_arraycopy
   681504261    0.58%       68  java.util.Arrays.copyOf
…